pyproject.toml
[project]
name = "cnn-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "python-multipart>=0.0.20",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "uvicorn[standard]>=0.38.0",
]
